# swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "MEDIA BASE URL : localhost:5000/images/ \n\n Authorization:-(In login Api Token Get)  ",
      "version": "1.0.0",
      "title": "Cricket App"
    },
    "host": "localhost:5000/",
    "basePath": "apis",
...
    "paths": {
...
      "/create_team": {
        "post": {
          "tags": ["Teams"],
          "summary": "Create Team",
          "operationId": "create_team",
          "produces": ["application/json"],
          "security": [
            {
              "APIKeyHeader": []
            }
          ],
          "consumes": ["multipart/form-data"],
          "parameters": [
              {
                  "name": "name",
                  "in": "formData",
                  "description": "Team Name",
                  "required": true,
                  "type": "string"
              },
              {
                "name": "city",
                "in": "formData",
                "description": "Team City",
                "required": true,
                "type": "string"
            },
            {
              "name": "team_photo",
              "in": "formData",
              "description": "Team Photo",
              "required": false,
              "type": "file"
            }
          ],
          "responses": {
            "200": {
              "description": "Logout  Successfully",
              "schema": {
                "type": "string"
              }
            },
            "400": {
              "description": "Error while logout"
            }
          }
        }
      },
...
      "/create_match": {
        "post": {
          "tags": ["Teams"],
          "summary": "Create match between teams",
          "operationId": "create_match",
          "produces": ["application/json"],
          "security": [
            {
              "APIKeyHeader": []
            }
          ],
          "consumes": ["multipart/form-data"],
          "parameters": [
              {
                  "name": "team1_id",
                  "in": "formData",
                  "description": "Enter  team 1 id",
                  "required": true,
                  "type": "string"
              },
              {
                "name": "team2_id",
                "in": "formData",
                "description": "Enter team 2 id",
                "required": true,
                "type": "string"
            },
            {
              "name": "total_over",
              "in": "formData",
              "description": "Enter total over of match",
              "required": true,
              "type": "string"
            },
            {
              "name": "city",
              "in": "formData",
              "description": "Enter city of match",
              "required": true,
              "type": "string"
            },
            {
              "name": "venue_ground",
              "in": "formData",
              "description": "Enter ground of match",
              "required": true,
              "type": "string"
            },
            {
              "name": "match_date",
              "in": "formData",
              "description": "Enter match date (format : YYYY-MM-DD)",
              "required": true,
              "type": "string"
            },
            {
              "name": "match_time",
              "in": "formData",
              "description": "Enter match time (in 24 hour format : 20:30- >like for 08:30)",
              "required": true,
              "type": "string"
            },
            {
              "name": "scorer_id",
              "in": "formData",
              "description": "Enter scorer id",
              "required": true,
              "type": "string"
            },
            {
              "name": "status",
              "in": "formData",
              "description": "Enter status (1.upcoming)",
              "required": true,
              "type": "string"
            },
            {
              "name": "team_one_json",
              "in": "formData",
              "description": "Enter team one json with there type (WK,BAT,BOWL,ALD)",
              "required": true,
              "type": "string"
            },
            {
              "name": "team_two_json",
              "in": "formData",
              "description": "Enter team two json with there type (WK,BAT,BOWL,ALD)",
              "required": true,
              "type": "string"
            },
            {
              "name": "player_list",
              "in": "formData",
              "description": "Player list array of both the team with 2 key team id and player id to add in scoreboard.",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "Logout  Successfully",
              "schema": {
                "type": "string"
              }
            },
            "400": {
              "description": "Error while logout"
            }
          }
        }
      },
...
    },
    "securityDefinitions": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    },
...
    "schemes": ["http"]
  }